+++ apps/web/src/views/Home/components/Banners/hooks/useMultipleBannerConfig.tsx
@@ -11,6 +11,7 @@ import { TopTraderBanner } from '../TopTraderBanner'
 import UserBanner from '../UserBanner'
 import { V4InfoBanner } from '../V4InfoBanner'
 import VeCakeBanner from '../VeCakeBanner'
+import { WBNBFixedStakingBanner } from '../WBNBFixedStakingBanner'
 import WebNotificationBanner from '../WebNotificationBanner'
 import useIsRenderCompetitionBanner from './useIsRenderCompetitionBanner'
 import useIsRenderIfoBanner from './useIsRenderIFOBanner'
@@ -46,6 +47,10 @@ export const useMultipleBannerConfig = () => {
         banner: <UserBanner />,
       },
       { shouldRender: isRenderIFOBanner || Boolean(countdown), banner: <NewIFOBanner /> },
+      {
+        shouldRender: true,
+        banner: <WBNBFixedStakingBanner />,
+      },
       {
         shouldRender: true,
         banner: <V4InfoBanner />,