+++ apps/web/src/views/Swap/MMLinkPools/hooks/useIsMMTradeBetter.tsx
@@ -1,25 +1,24 @@
 import { Currency, Trade, TradeType, ZERO } from '@pancakeswap/sdk'
 import { SmartRouterTrade } from '@pancakeswap/smart-router'
-import { LegacyTradeWithStableSwap as TradeWithStableSwap } from '@pancakeswap/smart-router/legacy-router'
 import { useMemo } from 'react'
 import { Field } from 'state/swap/actions'
 
-interface Options {
+interface Options<T> {
   independentField: Field
-  trade?: TradeWithStableSwap<Currency, Currency, TradeType> | SmartRouterTrade<TradeType> | null
-  v2Trade?: Trade<Currency, Currency, TradeType> | null
+  trade?: T | null
+  v2Trade?: Pick<Trade<Currency, Currency, TradeType>, 'inputAmount' | 'outputAmount'> | null
   tradeWithMM?: SmartRouterTrade<TradeType> | null
   isMMQuotingPair?: boolean
   isExpertMode?: boolean
 }
 
-export const useIsTradeWithMMBetter = ({
+export const useIsTradeWithMMBetter = <T extends Pick<SmartRouterTrade<TradeType>, 'inputAmount' | 'outputAmount'>>({
   independentField,
   trade,
   v2Trade,
   tradeWithMM,
   isExpertMode = false,
-}: Options) => {
+}: Options<T>) => {
   return useMemo(() => {
     const isExactIn = independentField === Field.INPUT
     if (