+++ packages/smart-router/evm/v3-router/utils/getPriceImpact.ts
@@ -1,9 +1,13 @@
 import { CurrencyAmount, Percent, TradeType } from '@pancakeswap/sdk'
 
-import { SmartRouterTrade } from '../types'
+import { Route, SmartRouterTrade } from '../types'
 import { getMidPrice } from './route'
 
-export function getPriceImpact(trade: SmartRouterTrade<TradeType>): Percent {
+export function getPriceImpact(
+  trade: Pick<SmartRouterTrade<TradeType>, 'outputAmount' | 'routes'> & {
+    routes: Pick<Route, 'path' | 'pools'>[]
+  },
+): Percent {
   let spotOutputAmount = CurrencyAmount.fromRawAmount(trade.outputAmount.currency.wrapped, 0)
   for (const route of trade.routes) {
     const { inputAmount } = route